refactor(1014): extract InsightMetricsUtils to src/analytics/insight_metrics_utils.py (P11, Cat E)
Move the InsightMetricsUtils class body (~328 LOC, 20+ static methods)
from MistHelper.py to a canonical module at
src/analytics/insight_metrics_utils.py. MistHelper.py keeps a top-level
re-export alias so historical MistHelper.InsightMetricsUtils / mh.
InsightMetricsUtils callers keep working.

Cross-class references (ConstDefinitionsExporter) and the module-level
apisession global are resolved lazily via
importlib.import_module("MistHelper") inside export_const_insight_metrics
to keep FR-028 IG-health clean (no top-level MistHelper import).

Coverage omit list updated for src/analytics/insight_metrics_utils.py.
FR-028 IG-health probe: PASS (MistHelper NOT in sys.modules after
importing src.analytics.insight_metrics_utils).

Initiative: #1014 (Cat E, position 11 of 24)
